Untested first pass at integrating chaining syntax
The process so far was to build up a "new" Unexpected, using only
method chaining, copying methods from the real Unexpected as they were
needed. This commit is the second stage, where the changed methods have
been copied _back_ to the real Unexpected.

This is untested, but it's a first pass.