Support forwarding flags in nested assertions
A small change was needed to the `to equal` assertion, as the `not` was
not previously a flag, but rather two separate assertions - this made
no difference with strings, but as only the flags can be passed to the
handler, it is not possible to switch handlers based on forwarded flags.

There may be other locations where this is an issue.