Split assertions when using `.shift`
When `expect.shift` is called from a handler running from a string
assertion, the assertion handler for the shifted subject is called
synchronously. When chaining, the next assertion is called when the
handler returns, using the new subject given in the `.shift` call.

This meant that the `UnexpectedError` aren't nested correctly.

When we catch the error, we now split out the assertions that were
called following a `.shift` call, and recreate the missing
`UnexpectedError` objects. This gets us close enough (we think) to
feature parity for shifting. There _could_ be some differences if a
handler performs some action following the shift, but there doesn't seem
to be a logical reason to do that. We may however, later discover
assertions of plugins that are doing some tricks with this.